+++ src/app/modules/leaderboard/leaderboard.component.ts
@@ -1,5 +1,4 @@
-import { CdkVirtualScrollViewport } from '@angular/cdk/scrolling'
-import { Component, TemplateRef, ViewChild } from '@angular/core'
+import { Component, TemplateRef } from '@angular/core'
 import { MatDialog } from '@angular/material'
 
 import { AuthService } from '@core/authentication/auth.service'
@@ -11,9 +10,6 @@ import { FirestoreService } from '@core/firestore.service'
   styleUrls: ['./leaderboard.component.scss']
 })
 export class LeaderboardComponent {
-  @ViewChild(CdkVirtualScrollViewport)
-  viewport: CdkVirtualScrollViewport
-
   aboutLeadDialog: any
 
   constructor(